+++ L107_028/interval.cpp
@@ -1,3 +1,7 @@
+// Мустафина Анна
+// pеализация функций intersects (функция для проверки пересечения интервалов) и ExhaustiveScheduling
+// (функция для нахождения наибольшего непересекающегося подмножества интервалов методом полного перебора)
+
 #include "interval.h"
 #include <vector>
 #include <algorithm>
