+++ L107_028/interval.h
@@ -1,3 +1,6 @@
+// Мустафина Анна
+// заголовочный файл с объявлением структуры Interval (структура для хранения интервалов)
+
 #ifndef INTERVAL_H
 #define INTERVAL_H
