+++ L108_029/Film.cpp
@@ -1,3 +1,8 @@
+// Мустафина Анна
+// реализация функций compareFilmsByEnd (сравнивает фильм по времени окончания съёмки)
+// doFilmsOverlap (проверяет пересечение съемок двух фильмов) и optimalScheduling 
+// (выбирает оптимальное расписание на основе жадного алгоритма) 
+
 #include "Film.h"
 #include <algorithm>
