+++ L108_029/Film.h
@@ -1,3 +1,6 @@
+// Мустафина Анна
+//заголовочный файл с объявлением структуры Film (структура для хранения информации о фильме)
+
 #ifndef FILM_H
 #define FILM_H
 #include <vector>
